* info/nodes.c (SUBFILE), info/nodes.h (TAG): go back to long type instead of size_t even if the variable should never be negative as it is more robust for comparisons, as discussed in: https://lists.gnu.org/archive/html/bug-texinfo/2024-10/msg00067.html

+++ info/nodes.c
@@ -356,7 +356,7 @@ get_tags_of_indirect_tags_table (FILE_BUFFER *file_buffer,
      an intermediate value. */
   typedef struct {
     char *filename;
-    size_t first_byte;
+    long first_byte;          /* Should never be negative */
   } SUBFILE;
 
   SUBFILE **subfiles = NULL;
